requirements.txt
absl-py==0.12.0
appnope==0.1.2
astroid==2.11.5
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
clr-loader==0.1.7
CSIKit==2.3
cycler==0.10.0
Cython==0.29.22
debugpy==1.5.1
decorator==5.1.0
dill==0.3.5.1
distlib==0.3.2
docopt==0.6.2
entrypoints==0.3
fastapi==0.63.0
filelock==3.0.12
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
h11==0.12.0
idna==2.10
install==1.3.5
ipykernel==6.6.0
ipython==7.30.1
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
llvmlite==0.37.0
Markdown==3.3.4
matplotlib==3.4.1
matplotlib-inline==0.1.3
mccabe==0.7.0
nest-asyncio==1.5.4
numba==0.54.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
pipreqs==0.4.11
platformdirs==2.2.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.1
Pygments==2.10.0
pylint==2.14.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
python-multipart==0.0.5
pythonnet==3.0.0a2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.6.2
six==1.16.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.4
uvicorn==0.13.4
virtualenv==20.7.0
virtualenv-clone==0.5.6
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.14.1
xgboost==1.5.1
yarg==0.1.9
